(#191): 지도 lazy 컴포넌트 적용
